librustc_incremental: Unconfigure tests during normal build
